feat: rewrite the PR name validator (#AB65381)
Instead of using amannn's third party action to perform the validation, we reuse the check done in the commit validation action and simply check the prefix of the PR title.

We also fail the check if the title is invalid without working with any labels.

The set of permissions needed is now reduced to

permissions: {}

It previously required

permission:
  contents: read
  pull-requests: write

Clients can safely remove the label invalid_PR_name from their repos after upgrading.