chore: fix vulnerability in PR name validator (#AB66587)
This fixes an injection vulnerability that allows an attacker to run arbitrary shell commands by manipulating the PR title. Quotes in the if statement are not part of this fix.

Anyone creating a PR could craft a title like `feat: not a malicious title at all $((ls ~))` and have it run. In this case, simply listing the contents of the runner user's home dir in the action log.